Add endpoint to re-index a specific record
Can hit `/establishments/:id` `/projects/:id` or `/profiles/:id` to refresh the index for a particular record without needing to rebuild the entire index.

The eventual use-case for this will be a workflow hook that refreshes the index on those entities whenever a task resolves that would update the data.