Split search term into words and combine with AND
If a user enters a search term with multiple words then run a separate search for each word and require that all words match in a result.

This means that adding additional search terms will make a search more specific rather than less specific, which aligns better with user expectations.