Increase max results window setting for profiles index
This allows the complete set of profile data from prod to be included in a search. It can result in an increase increase in resource usage and impaired performance, but testing locally on a ~30,000 record dataset with a less powerful elastic instance than prod works fine - presumably because the index data is still small (~5mb total size).

Over 10,000 results the results counter loses accuracy and only returns a count of `gte` 10000, so if that happens then reset the result counter to assume we're getting all the results, because in 100% of our use cases that's what is happening.