Set search type when no sort order is defined
Doing relevancy scoring on small indexes doesn't work reliably by default across multiple shards. This means search results can be returned in an unexpected order when sorting by score (default sort).

The workaround for this is to set the search_type property so all shards are queried when assembling statisitics for relevancy scores. Since this has a perf impact but is only relevant when sorting by score set the param only if no specific sort column is defined.