Move transform stream creation into index function
Previously the transform stream for document transforms was only being created as a single instance at startup, which meant that once it was used once it remained closed and no more results would be processed. This meant that only the first index after a process restart would find results.

Move the stream instance creation into the index function so a fresh open stream is created each time.